move play command to own file
