+++ commands/ping.js
@@ -1,6 +1,6 @@
 module.exports = {
   name: 'ping',
-  description: 'Ping the bot',
+  description: 'Pong!',
   aliases: [],
   command: (message, arguments, serverQueue, client) => {
     return message.channel.send('pong');
